s3:include: change lock_struct->fnum to uint64_t
Note: this changes the format of brlock.tdb!

Pair-Programmed-With: Stefan Metzmacher <metze@samba.org>

Autobuild-User: Michael Adam <obnox@samba.org>
Autobuild-Date: Wed Jun  6 23:22:00 CEST 2012 on sn-devel-104

+++ source3/include/locking.h
@@ -66,7 +66,7 @@ struct lock_struct {
 	struct lock_context context;
 	br_off start;
 	br_off size;
-	uint16 fnum;
+	uint64_t fnum;
 	enum brl_type lock_type;
 	enum brl_flavour lock_flav;
 };
